requirements.txt
Pillow==5.0.0
beautifulsoup4==4.6.0